+++ src/commands/__init__.py
@@ -1,0 +1,6 @@
+from .base import BaseCommand  # noqa
+from .card import CardCommand  # noqa
+from .faq import FaqCommand  # noqa
+from .deck import DeckCommand  # noqa
+from .cimg import CimgCommand  # noqa
+from .overlap import OverlapCommand  # noqa